created testimonials section, moved home sections to home folder
