+++ src/app/about/page.tsx
@@ -2,13 +2,13 @@
 import { useEffect, useState } from "react";
 import RoleButton from "../../components/RoleButton";
 import TeamCard from "../../components/TeamCard";
-import {teams} from "./teamData";
-import data from "./data.json"
+import { teams } from "./teamData";
+import data from "./data.json";
+import Testimonials from "@/sections/about/Testimonials";
 
 export default function Page() {
-
-  const [activeButton, setActiveButton] = useState<string | null> (null);
-  const [isDesktop, setIsDesktop] = useState<boolean | null> (null);
+  const [activeButton, setActiveButton] = useState<string | null>(null);
+  const [isDesktop, setIsDesktop] = useState<boolean | null>(null);
 
   useEffect(() => {
     const handleResize = () => {
@@ -21,9 +21,9 @@ export default function Page() {
     const isDesktopValue = newWidth > 1024;
     setIsDesktop(isDesktopValue);
 
-    const selectedTeam = localStorage.getItem("selectedTeam") || "Directors"
-    setActiveButton(selectedTeam)
-    window.innerWidth > 1024? setIsDesktop(true): setIsDesktop(false)
+    const selectedTeam = localStorage.getItem("selectedTeam") || "Directors";
+    setActiveButton(selectedTeam);
+    window.innerWidth > 1024 ? setIsDesktop(true) : setIsDesktop(false);
 
     // Add event listener to handle window resize
     window.addEventListener("resize", handleResize);
@@ -31,14 +31,17 @@ export default function Page() {
     return () => {
       window.removeEventListener("resize", handleResize);
     };
-  }, ); 
+  });
 
   const changeTeam = (buttonId: string) => {
     setActiveButton(buttonId);
     localStorage.setItem("selectedTeam", buttonId);
-    isDesktop!?document.getElementById("team")?.scrollIntoView({behavior:"smooth", block:"start"}): null;
+    isDesktop!
+      ? document
+          .getElementById("team")
+          ?.scrollIntoView({ behavior: "smooth", block: "start" })
+      : null;
   };
-
 
   return (
     <main className="flex flex-col gap-28">
@@ -56,12 +59,16 @@ export default function Page() {
         </div>
       </section>
 
-      <section id="team" className=" pt-24 flex flex-col lg:flex-row gap-12 relative text-pretty">
-
+      <section
+        id="team"
+        className=" pt-24 flex flex-col lg:flex-row gap-12 relative text-pretty"
+      >
         <div className="w-full lg:w-[40%]">
           <div className=" flex gap-8 flex-col lg:sticky top-24 ">
             <h2 className="text-caption">MEET THE TEAM</h2>
-            <h3 className="title-1 emphasized">We’re all super cracked here at Surge</h3>
+            <h3 className="title-1 emphasized">
+              We’re all super cracked here at Surge
+            </h3>
             <p className="paragraph ">
               The organizers of Simon Fraser University’s largest hackathon, SFU
               Surge provides an inclusive space for students to explore...
@@ -73,10 +80,11 @@ export default function Page() {
                     buttonFunction={() => changeTeam(team.name)}
                     key={team.id}
                     title={team.name}
-                    icon = {team.icon}
-
+                    icon={team.icon}
                     className={`${
-                      activeButton === team.name ? "bg-brand-primary/30 text-text-brand-primary " : "bg-gray-800/10 text-text-secondary hover:bg-brand-primary-hover/20"
+                      activeButton === team.name
+                        ? "bg-brand-primary/30 text-text-brand-primary "
+                        : "bg-gray-800/10 text-text-secondary hover:bg-brand-primary-hover/20"
                     }`}
                   />
                 );
@@ -84,27 +92,28 @@ export default function Page() {
             </div>
           </div>
         </div>
-        <div className=" flex flex-col w-full lg:w-[65%] lg:ml-8 rounded-2xl
-       ">
+        <div
+          className=" flex flex-col w-full lg:w-[65%] lg:ml-8 rounded-2xl
+       "
+        >
           <div className=" flex flex-col  bg-[#121318]/80  p-5 rounded-2xl  gap-5 transition-all ">
-
-          {data.map((member) =>
-            activeButton === member.team ? (
-              <TeamCard
-                key={member.id}
-                name={member.name}
-                major={member.major}
-                role={member.position}
-                src = {member.src}
-                fallbackSrc="/headshots/placeholder.png"
-              />
-            ) : null
-          )}
+            {data.map((member) =>
+              activeButton === member.team ? (
+                <TeamCard
+                  key={member.id}
+                  name={member.name}
+                  major={member.major}
+                  role={member.position}
+                  src={member.src}
+                  fallbackSrc="/headshots/placeholder.png"
+                />
+              ) : null
+            )}
           </div>
-         
         </div>
       </section>
 
+      <Testimonials />
     </main>
   );
 }
