+++ src/app/page.tsx
@@ -1,6 +1,6 @@
-import Landing from "../sections/Landing"
-import Events from "../sections/Events"
-import Involvement from "../sections/Involvement"
+import Landing from "../sections/home/Landing";
+import Events from "../sections/home/Events";
+import Involvement from "../sections/home/Involvement";
 
 export default function Home() {
   return (
