+++ src/sections/about/Testimonials.tsx
@@ -1,0 +1,3 @@
+export default function Testimonials() {
+  return <></>;
+}
